src/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
ffmpeg-python==0.2.0
youtube_dl==2021.12.17